# Fix massive performance issue in read_to_end