@alexcrichton 2015-03-29
Very nice wins! Could you add some comments about the rationale of the growing behavior to the `read_to_end` function? Other than that r=me