@alexcrichton 2015-03-29
@bors: retry